changed method`s parameters (for 'id': int => long)

+++ src/main/java/dclvs/moviehubui/repositories/UserRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findUserById(int id);
+    Optional<User> findUserById(Long id);
 
     List<User> findAllUsers();
